Allow directory services to be configured on different addresses

+++ packages/aserto/src/aserto/client/directory/v3/__init__.py
@@ -1,5 +1,5 @@
 import datetime
-import typing
+import typing 
 
 import aserto.directory.exporter.v3 as exporter
 import aserto.directory.importer.v3 as importer
@@ -26,24 +26,71 @@ from aserto.client.directory.v3.helpers import (
     RelationsResponse,
 )
 
+def build_grpc_channel(address: str, default_address: str, ca_cert_path: str) -> typing.Any:
+    if address == "" and default_address == "":
+        return None
+        
+    if address == "":
+        return grpc.secure_channel(
+        target=default_address, credentials=directory.channel_credentials(cert=ca_cert_path)
+    )
+        
+    return grpc.secure_channel(
+        target=address, credentials=directory.channel_credentials(cert=ca_cert_path)
+    )
 
 class Directory:
     def __init__(
         self,
-        address: str,
         api_key: str = "",
         tenant_id: str = "",
         ca_cert_path: str = "",
+        address: str = "",
+        reader_address: str = "",
+        writer_address: str = "",
+        importer_address: str = "",
+        exporter_address: str = "",
+        model_address: str = "",
     ) -> None:
-        self._channel = grpc.secure_channel(
-            target=address, credentials=directory.channel_credentials(cert=ca_cert_path)
-        )
+        
+        validation_error = directory.validate_addresses(address=address, reader_address=reader_address, writer_address=writer_address,
+                              importer_address=importer_address, exporter_address=exporter_address, model_address=model_address)
+        if validation_error is not None:
+            raise validation_error
+
+        self._reader_channel = build_grpc_channel(reader_address, address, ca_cert_path)
+        self._writer_channel = build_grpc_channel(writer_address, address, ca_cert_path)
+        self._importer_channel = build_grpc_channel(importer_address, address, ca_cert_path)
+        self._exporter_channel = build_grpc_channel(exporter_address, address, ca_cert_path)
+        self._model_channel = build_grpc_channel(model_address, address, ca_cert_path)
+
         self._metadata = directory.get_metadata(api_key=api_key, tenant_id=tenant_id)
-        self.reader = reader.ReaderStub(self._channel)
-        self.writer = writer.WriterStub(self._channel)
-        self.model = model.ModelStub(self._channel)
-        self.importer = importer.ImporterStub(self._channel)
-        self.exporter = exporter.ExporterStub(self._channel)
+        self.reader = (
+            reader.ReaderStub(self._reader_channel)
+            if self._reader_channel is not None
+            else None
+        )
+        self.writer = (
+            writer.WriterStub(self._writer_channel)
+            if self._writer_channel is not None
+            else None
+        )
+        self.model = (
+            model.ModelStub(self._model_channel)
+            if self._model_channel is not None
+            else None
+        )
+        self.importer = (
+            importer.ImporterStub(self._importer_channel)
+            if self._importer_channel is not None
+            else None
+        )
+        self.exporter = (
+            exporter.ExporterStub(self._exporter_channel)
+            if self._exporter_channel is not None
+            else None
+        )
+
 
     @typing.overload
     def get_object(
@@ -91,6 +138,9 @@ class Directory:
         a directory object or, if with_relations is True, a GetObjectResponse.
         """
 
+        if self.reader is None:
+            raise directory.NilClient
+
         try:
             response = self.reader.GetObject(
                 reader.GetObjectRequest(
@@ -130,6 +180,9 @@ class Directory:
             list of directory objects
         """
 
+        if self.reader is None:
+            raise directory.NilClient
+
         try:
             response = self.reader.GetObjectMany(
                 reader.GetObjectManyRequest(param=(i.proto for i in identifiers)),
@@ -161,6 +214,8 @@ class Directory:
             page : PaginationResponse
                 the next page's token if there are more results
         """
+        if self.reader is None:
+            raise directory.NilClient
 
         response = self.reader.GetObjects(
             reader.GetObjectsRequest(object_type=object_type, page=page),
@@ -228,6 +283,10 @@ class Directory:
         properties: typing.Optional[typing.Union[typing.Mapping[str, typing.Any], Struct]] = None,
         etag: str = "",
     ) -> Object:
+        
+        if self.writer is None:
+            raise directory.NilClient
+
         obj = object
         if obj is None:
             props = (
@@ -270,6 +329,9 @@ class Directory:
         ----
         None
         """
+
+        if self.writer is None:
+            raise directory.NilClient
 
         self.writer.DeleteObject(
             writer.DeleteObjectRequest(
@@ -341,6 +403,9 @@ class Directory:
         OR
             a RelationResponse if with_objects is set to True
         """
+
+        if self.reader is None:
+            raise directory.NilClient
 
         try:
             response = self.reader.GetRelation(
@@ -415,6 +480,9 @@ class Directory:
                 retrieved page information — the size of the page, and the next page's token
         """
 
+        if self.reader is None:
+            raise directory.NilClient
+
         response = self.reader.GetRelations(
             reader.GetRelationsRequest(
                 object_type=object_type,
@@ -466,6 +534,9 @@ class Directory:
         The created relation
         """
 
+        if self.writer is None:
+            raise directory.NilClient
+
         response = self.writer.SetRelation(
             writer.SetRelationRequest(
                 relation=Relation(
@@ -512,6 +583,9 @@ class Directory:
         None
         """
 
+        if self.writer is None:
+            raise directory.NilClient
+
         self.writer.DeleteRelation(
             writer.DeleteRelationRequest(
                 object_type=object_type,
@@ -552,6 +626,9 @@ class Directory:
         ----
         True or False
         """
+        if self.reader is None:
+            raise directory.NilClient
+        
         response = self.reader.Check(
             reader.CheckRequest(
                 object_type=object_type,
@@ -591,6 +668,9 @@ class Directory:
         ----
         True or False
         """
+
+        if self.reader is None:
+            raise directory.NilClient
 
         response = self.reader.CheckRelation(
             reader.CheckRelationRequest(
@@ -633,6 +713,9 @@ class Directory:
         True or False
         """
 
+        if self.reader is None:
+            raise directory.NilClient
+
         response = self.reader.CheckPermission(
             reader.CheckPermissionRequest(
                 object_type=object_type,
@@ -666,6 +749,10 @@ class Directory:
         ----
         The current manifest or None.
         """
+
+        if self.model is None:
+            raise directory.NilClient
+
         headers = self._metadata
         if etag:
             headers += (("if-none-match", etag),)
@@ -699,6 +786,9 @@ class Directory:
         None
         """
 
+        if self.model is None:
+            raise directory.NilClient
+
         headers = self._metadata
         if etag:
             headers += (("if-match", etag),)
@@ -731,6 +821,9 @@ class Directory:
         ImportResponse:
             a summary of the total number of object and relations imported.
         """
+
+        if self.importer is None:
+            raise directory.NilClient
 
         def _import_iter() -> typing.Iterator[importer.ImportRequest]:
             for item in data:
@@ -771,6 +864,10 @@ class Directory:
         start_from: typing.Optional[datetime.datetime]
             if provided, only objects and relations that have been modified after this date are exported.
         """
+
+        if self.exporter is None:
+            raise directory.NilClient
+
         req = exporter.ExportRequest(options=options)
         if start_from is not None:
             req.start_from.FromDatetime(dt=start_from)
@@ -784,8 +881,20 @@ class Directory:
 
     def close(self) -> None:
         """Closes the gRPC channel"""
+        if self._reader_channel is not None:
+            self._reader_channel.close()
 
-        self._channel.close()
+        if self._writer_channel is not None:
+            self._writer_channel.close()
+
+        if self._importer_channel is not None:
+            self._importer_channel.close()
+
+        if self._exporter_channel is not None:
+            self._exporter_channel.close()
+        
+        if self._model_channel is not None:
+            self._model_channel.close()
 
     def __enter__(self):
         return self
